+++ lib/src/layout_overlays.dart
@@ -165,10 +165,10 @@ class _OverlayBuilderState extends State<OverlayBuilder> {
   }
 
   void addToOverlay(OverlayEntry overlayEntry) async {
+    final showCaseContext = ShowCaseWidget.of(context).context;
     if (mounted) {
-      if (ShowCaseWidget.of(context)?.context != null &&
-          Overlay.of(ShowCaseWidget.of(context)!.context) != null) {
-        Overlay.of(ShowCaseWidget.of(context)!.context)!.insert(overlayEntry);
+      if (Overlay.of(showCaseContext) != null) {
+        Overlay.of(showCaseContext)!.insert(overlayEntry);
       } else {
         if (Overlay.of(context) != null) {
           Overlay.of(context)!.insert(overlayEntry);